Replace fmt with logging. Remove fatal logging with info logging and 500 status codes.
