+++ main.go
@@ -15,39 +15,149 @@ import (
 	_ "github.com/lib/pq"
 )
 
-// local host
-// const (
-// 	host     = "localhost"
-// 	port     = 5432
-// 	user     = "zini"
-// 	password = "zini"
-// 	dbname   = "perfsearch"
-// 	sslmode  = "disable"
-// )
-
-// heroku server
-const (
-	host     = "ec2-184-72-236-57.compute-1.amazonaws.com"
-	port     = 5432
-	user     = "wrwcqifhvfkkjw"
-	password = "08f2594b47185df91e8bd0513405b8f5f4831089dc59c27387cd89e465b96015"
-	dbname   = "d3hpcrkvokd5i"
-	sslmode  = "require"
-)
-
 type idList struct {
 	IdList []int64 `json:"ids" binding:"required"`
 }
 
-type Detail struct {
-	Title  string
-	Studio string
-	Price  string
-	Rating string
-	Year   string
-	Genre  string
-	Upc    string
-	ID     int64
+func search(c *gin.Context) {
+	// get search keywords
+	keyword := c.Query("keyword")
+	log.Printf("keyword is: %s\n", keyword)
+
+	var dvds []Detail
+
+	// connect to database
+	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
+		"password=%s dbname=%s sslmode=%s",
+		host, port, user, password, dbname, sslmode)
+	db, err := sql.Open("postgres", psqlInfo)
+	if err != nil {
+		log.Println(err)
+		c.JSON(500, "")
+		return
+	}
+	log.Println("Successfully connected to DB")
+
+	sqlStatement := `SELECT * FROM dvds WHERE LOWER(title) LIKE '%' || $1 || '%' ;`
+	log.Println("'%" + keyword + "%'")
+	rows, err := db.Query(sqlStatement, keyword)
+
+	if err != nil {
+		log.Println(err)
+		c.JSON(500, "")
+		return
+	}
+
+	for rows.Next() {
+		var dvd Detail
+		queryErr := rows.Scan(&dvd.Title, &dvd.Studio, &dvd.Price, &dvd.Rating, &dvd.Year,
+			&dvd.Genre, &dvd.Upc, &dvd.ID)
+		switch queryErr {
+		case sql.ErrNoRows:
+			log.Println("No rows were returned!")
+			return
+		case nil:
+			dvds = append(dvds, dvd)
+			log.Println(dvd)
+		default:
+			panic(queryErr)
+		}
+	}
+
+	defer db.Close()
+	log.Println(dvds)
+	c.JSON(200, dvds)
+}
+
+func getMoviesByIDs(c *gin.Context) {
+	// get query ids
+	ids := idList{}
+	var idList []int64
+	var dvds []Detail
+	
+	// This reads c.Request.Body and stores the result into the context.
+	if err := c.ShouldBindBodyWith(&ids, binding.JSON); err == nil {
+		idList = ids.IdList
+		log.Println(idList)
+	}
+
+	// connect to database
+	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
+		"password=%s dbname=%s sslmode=%s",
+		host, port, user, password, dbname, sslmode)
+	db, err := sql.Open("postgres", psqlInfo)
+	if err != nil {
+		log.Println(err)
+		c.JSON(500, "")
+		return
+	}
+	log.Println("Successfully connected to DB")
+
+	sqlStatement := `SELECT * FROM dvds WHERE id = any($1);`
+	rows, err := db.Query(sqlStatement, pq.Array(idList))
+
+	if err != nil {
+		log.Println(err)
+		c.JSON(500, "")
+		return
+	}
+
+	for rows.Next() {
+		var dvd Detail
+		queryErr := rows.Scan(&dvd.Title, &dvd.Studio, &dvd.Price, &dvd.Rating, &dvd.Year,
+			&dvd.Genre, &dvd.Upc, &dvd.ID)
+		switch queryErr {
+		case sql.ErrNoRows:
+			log.Println("No rows were returned!")
+			return
+		case nil:
+			dvds = append(dvds, dvd)
+			log.Println(dvd)
+		default:
+			panic(queryErr)
+		}
+	}
+
+	defer db.Close()
+	log.Println(dvds)
+	c.JSON(200, dvds)
+}
+
+func getMovieByID(c *gin.Context) {
+	id, err := strconv.ParseInt(c.Query("id"), 10, 64)
+
+	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
+		"password=%s dbname=%s sslmode=%s",
+		host, port, user, password, dbname, sslmode)
+	db, err := sql.Open("postgres", psqlInfo)
+	if err != nil {
+		log.Println(err)
+		c.JSON(500, "")
+		return
+	}
+	log.Println("Successfully connected to DB")
+
+	sqlStatement := `SELECT * FROM dvds WHERE id=$1;`
+	var dvd Detail
+	row := db.QueryRow(sqlStatement, id)
+
+	queryErr := row.Scan(&dvd.Title, &dvd.Studio, &dvd.Price, &dvd.Rating, &dvd.Year,
+		&dvd.Genre, &dvd.Upc, &dvd.ID)
+
+	log.Println(dvd)
+
+	switch queryErr {
+	case sql.ErrNoRows:
+		log.Println("No rows were returned!")
+		return
+	case nil:
+		log.Println(dvd)
+	default:
+		panic(queryErr)
+	}
+
+	defer db.Close()
+	c.JSON(200, dvd)
 }
 
 func main() {
@@ -70,134 +180,4 @@ func main() {
 	router.POST("/getMoviesByIds", getMoviesByIDs)
 
 	router.Run(":" + port)
-}
-
-func search(c *gin.Context) {
-	// get search keywords
-	keyword := c.Query("keyword")
-	fmt.Printf("keyword is: %s\n", keyword)
-
-	dvds := []Detail{}
-
-	// connect to database
-	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
-		"password=%s dbname=%s sslmode=%s",
-		host, port, user, password, dbname, sslmode)
-	db, err := sql.Open("postgres", psqlInfo)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Successfully connected to DB")
-
-	sqlStatement := `SELECT * FROM dvds WHERE LOWER(title) LIKE '%' || $1 || '%' ;`
-	fmt.Println("'%" + keyword + "%'")
-	rows, err := db.Query(sqlStatement, keyword)
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	for rows.Next() {
-		var dvd Detail
-		queryErr := rows.Scan(&dvd.Title, &dvd.Studio, &dvd.Price, &dvd.Rating, &dvd.Year,
-			&dvd.Genre, &dvd.Upc, &dvd.ID)
-		switch queryErr {
-		case sql.ErrNoRows:
-			fmt.Println("No rows were returned!")
-			return
-		case nil:
-			dvds = append(dvds, dvd)
-			fmt.Println(dvd)
-		default:
-			panic(queryErr)
-		}
-	}
-
-	defer db.Close()
-	fmt.Println(dvds)
-	c.JSON(200, dvds)
-}
-
-func getMoviesByIDs(c *gin.Context) {
-	// get query ids
-	ids := idList{}
-	idList := []int64{}
-	dvds := []Detail{}
-	// This reads c.Request.Body and stores the result into the context.
-	if err := c.ShouldBindBodyWith(&ids, binding.JSON); err == nil {
-		idList = ids.IdList
-		fmt.Println(idList)
-	}
-
-	// connect to database
-	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
-		"password=%s dbname=%s sslmode=%s",
-		host, port, user, password, dbname, sslmode)
-	db, err := sql.Open("postgres", psqlInfo)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Successfully connected to DB")
-
-	sqlStatement := `SELECT * FROM dvds WHERE id = any($1);`
-	rows, err := db.Query(sqlStatement, pq.Array(idList))
-
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	for rows.Next() {
-		var dvd Detail
-		queryErr := rows.Scan(&dvd.Title, &dvd.Studio, &dvd.Price, &dvd.Rating, &dvd.Year,
-			&dvd.Genre, &dvd.Upc, &dvd.ID)
-		switch queryErr {
-		case sql.ErrNoRows:
-			fmt.Println("No rows were returned!")
-			return
-		case nil:
-			dvds = append(dvds, dvd)
-			fmt.Println(dvd)
-		default:
-			panic(queryErr)
-		}
-	}
-
-	defer db.Close()
-	fmt.Println(dvds)
-	c.JSON(200, dvds)
-}
-
-func getMovieByID(c *gin.Context) {
-	id, err := strconv.ParseInt(c.Query("id"), 10, 64)
-
-	psqlInfo := fmt.Sprintf("host=%s port=%d user=%s "+
-		"password=%s dbname=%s sslmode=%s",
-		host, port, user, password, dbname, sslmode)
-	db, err := sql.Open("postgres", psqlInfo)
-	if err != nil {
-		log.Fatal(err)
-	}
-	fmt.Println("Successfully connected to DB")
-
-	sqlStatement := `SELECT * FROM dvds WHERE id=$1;`
-	var dvd Detail
-	row := db.QueryRow(sqlStatement, id)
-
-	queryErr := row.Scan(&dvd.Title, &dvd.Studio, &dvd.Price, &dvd.Rating, &dvd.Year,
-		&dvd.Genre, &dvd.Upc, &dvd.ID)
-
-	fmt.Println(dvd)
-
-	switch queryErr {
-	case sql.ErrNoRows:
-		fmt.Println("No rows were returned!")
-		return
-	case nil:
-		fmt.Println(dvd)
-	default:
-		panic(queryErr)
-	}
-
-	defer db.Close()
-	c.JSON(200, dvd)
 }
